requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
alabaster==0.7.12
apturl==0.5.2
async-timeout==4.0.2
asyncio==3.4.3
attrs==22.2.0
Babel==2.8.0
bcrypt==3.2.0
beautifulsoup4==4.11.1
bigquery-schema-generator==1.5
blinker==1.4
Brlapi==0.8.3
CacheControl==0.12.11
cachetools==4.2.4
certifi==2020.6.20
chardet==4.0.0
charset-normalizer==2.1.1
click==8.0.3
colorama==0.4.4
command-not-found==0.3
cryptography==3.4.8
cssselect==1.1.0
cssutils==2.6.0
cupshelpers==1.0
curlify==2.2.1
db-dtypes==1.0.4
dbus-python==1.2.18
decorator==5.1.1
defer==1.0.6
distro==1.7.0
distro-info===1.1build1
dnspython==2.2.1
docopt==0.6.2
docutils==0.17.1
duplicity==0.8.21
facebook-business==15.0.0
fasteners==0.14.1
feedparser==6.0.8
firebase-admin==5.4.0
Flask==2.2.2
Flask-Cors==3.0.10
Flask-Minify==0.41
Flask-SQLAlchemy==3.0.2
frozenlist==1.3.3
future==0.18.2
gitsome==0.8.0
google-api-core==2.10.1
google-api-python-client==2.63.0
google-auth==1.35.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.5.3
google-cloud==0.34.0
google-cloud-appengine-logging==1.1.4
google-cloud-audit-log==0.2.4
google-cloud-bigquery==3.3.3
google-cloud-bigquery-storage==2.16.1
google-cloud-core==2.3.2
google-cloud-firestore==2.7.0
google-cloud-logging==3.2.3
google-cloud-pubsub==2.13.10
google-cloud-secret-manager==2.12.6
google-cloud-storage==2.5.0
google-crc32c==1.5.0
google-resumable-media==2.4.0
googleapis-common-protos==1.56.4
greenlet==2.0.1
grpc-google-iam-v1==0.12.4
grpcio==1.49.1
grpcio-status==1.49.1
gspread==5.7.2
gyp==0.1
htmlmin==0.1.12
httplib2==0.20.2
idna==3.3
imagesize==1.3.0
importlib-metadata==4.6.4
itsdangerous==2.1.2
jeepney==0.7.1
Jinja2==3.0.3
jsmin==3.0.1
keyring==23.5.0
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lesscpy==0.15.1
lockfile==0.12.2
louis==3.20.0
lxml==4.9.1
macaroonbakery==1.3.1
Mako==1.1.3
MarkupSafe==2.1.1
monotonic==1.6
more-itertools==8.10.0
msgpack==1.0.4
multidict==6.0.3
netifaces==0.11.0
numpy==1.23.3
numpydoc==1.2
oauthlib==3.2.0
olefile==0.46
packaging==21.3
pandas==1.5.0
pandas-gbq==0.17.9
paramiko==2.9.3
pexpect==4.8.0
Pillow==9.0.1
ply==3.11
premailer==3.10.0
prompt-toolkit==3.0.28
proto-plus==1.22.1
protobuf==4.21.7
ptyprocess==0.7.0
pyarrow==9.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.20.1
pycountry==22.3.5
pycryptodome==3.15.0
pycups==2.0.1
pydata-google-auth==1.4.0
pygbq==0.26
Pygments==2.11.2
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
pymongo==4.2.0
PyNaCl==1.5.0
pyparsing==2.4.7
PyPDF2==2.11.1
pyRFC3339==1.1
python-apt==2.3.0+ubuntu2.1
python-dateutil==2.8.1
python-debian===0.1.43ubuntu1
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
rcssmin==1.1.1
reportlab==3.6.8
requests==2.25.1
requests-oauthlib==1.3.1
roman==3.3
rsa==4.9
screen-resolution-extra==0.0.0
searchconsole @ git+https://github.com/joshcarty/google-searchconsole@12c5cf5d86b6c4650055a2aeb6a7a59e5d5a3a7d
SecretStorage==3.3.1
setproctitle==1.2.2
sgmllib3k==1.0.0
six==1.16.0
snowballstemmer==2.2.0
soupsieve==2.3.2.post1
Sphinx==4.3.2
SQLAlchemy==1.4.44
systemd-python==234
tqdm==4.64.1
ubuntu-advantage-tools==27.11.2
ubuntu-drivers-common==0.0.0
ufw==0.36.1
unattended-upgrades==0.1
uritemplate==4.1.1
urllib3==1.26.5
usb-creator==0.3.7
validators==0.20.0
wadllib==1.3.6
wcwidth==0.2.5
Werkzeug==2.2.2
xdg==5
xkit==0.0.0
xonsh==0.11.0
xxhash==3.1.0
yagmail==0.15.293
yarl==1.8.2
zipp==1.0.0
